booktracker/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-allauth==0.63.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
requests==2.32.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
django-extensions==3.2.3
django-cors-headers==4.3.1
drf-yasg==1.21.7